Moved the RegRegistry class from the cachingreg to the dispatch module.
